Added changes of entity

+++ employee-management-hrms/employee-management-hrms/src/main/java/com/employee/management/hrms/repository/UserRepository.java
@@ -1,21 +1,15 @@
 package com.employee.management.hrms.repository;
 
-
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.employee.management.hrms.entity.User;
-
-
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
-    // Find user by email (for login)
     Optional<User> findByEmail(String email);
 
-    // Check if email already exists
     boolean existsByEmail(String email);
 }
-
